fix(logs): skip writing logs when opening the logs file fails
